chore: remove unused Dockerfiles, docker-compose, and related artifacts; add serverless support for backend
