--- backend/.dockerignore
@@ -1,5 +1,0 @@
-node_modules
-dist
-.env
-.env.*
-.git
