+++ backend/package.json
@@ -17,6 +17,7 @@
     "@prisma/adapter-pg": "^7.4.2",
     "@prisma/client": "^7.8.0",
     "bcrypt": "^6.0.0",
+    "serverless-http": "^2.8.0",
     "cors": "^2.8.6",
     "dotenv": "^17.3.1",
     "express": "^5.2.1",
