+++ backend/src/controllers/badgeController.ts
@@ -102,7 +102,7 @@ export const createBadge = async (req: Request, res: Response): Promise<void> =>
  */
 export const updateBadge = async (req: AuthRequest, res: Response): Promise<void> => {
   try {
-    const { id } = req.params;
+    const id = req.params.id as string;
     const { nama_badge, deskripsi, jenis_syarat, nilai_syarat } = req.body;
 
     const existing = await prisma.badges.findUnique({
@@ -148,7 +148,7 @@ export const updateBadge = async (req: AuthRequest, res: Response): Promise<void
  */
 export const deleteBadge = async (req: AuthRequest, res: Response): Promise<void> => {
   try {
-    const { id } = req.params;
+    const id = req.params.id as string;
 
     // Check if any users have this badge
     const earnedCount = await prisma.user_badges.count({
