+++ backend/vercel.json
@@ -1,0 +1,5 @@
+{
+  "version": 3,
+  "builds": [{ "src": "api/index.ts", "use": "@vercel/node" }],
+  "routes": [{ "src": "^/(.*)$", "dest": "/api/index.ts" }]
+}
